pyproject.toml
[project]
name = "ai-literature-assistant"
version = "0.1.0"
description = "AI-powered academic literature assistant for systematic review workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.40",
    "anthropic>=0.40",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "tenacity>=9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
    "ruff>=0.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
